requirements.txt
aiortc
fastapi
uvicorn
websockets
coqui-tts
numpy
scipy
whisperx
python-multipart
opuslib
mutagen
torch
torchaudio
